+++ src/main/webapp/resources/js/apis/analysis/analysis.js
@@ -279,6 +279,28 @@ export async function getAnalysisProvenanceByFile(submissionId, filename) {
     return { error };
   }
 }
+
+/**
+ * Gets the parsed excel data
+ * @return {Promise<*>} `data` contains the OK response; `error` contains error information if an error occurred.
+ */
+export async function parseExcel(submissionId, filename, sheetIndex) {
+  try {
+    const { data } = await axios.get(
+      `${ANALYSIS_URL}/${submissionId}/parseExcel`,
+      {
+        params: {
+          filename,
+          sheetIndex
+        }
+      }
+    );
+    return { data };
+  } catch (error) {
+    return { error };
+  }
+}
+
 /**
  * Get all single sample analysis output file info for the principal user.
  * @return {Promise<*>} `data` contains the OK response; `error` contains error information if an error occurred.